Invoke ESLint from the root
