+++ package.json
@@ -4,7 +4,12 @@
     "workspaces": [
         "packages/*"
     ],
-    "dependencies": {
-        "prettier": "^2.8.3"
+    "devDependencies": {
+        "@typescript-eslint/eslint-plugin": "^5.49.0",
+        "@typescript-eslint/parser": "^5.50.0",
+        "eslint": "^8.33.0",
+        "eslint-config-prettier": "^8.6.0",
+        "prettier": "^2.8.3",
+        "typescript": "^4.9.4"
     }
 }
