+++ packages/option-t/package.json
@@ -55,11 +55,7 @@
     "@babel/core": "^7.20.12",
     "@babel/plugin-syntax-typescript": "^7.20.0",
     "@babel/preset-env": "^7.20.2",
-    "@typescript-eslint/eslint-plugin": "^5.49.0",
-    "@typescript-eslint/parser": "^5.50.0",
     "del-cli": "^5.0.0",
-    "eslint": "^8.33.0",
-    "eslint-config-prettier": "^8.6.0",
     "picomatch": "^2.3.1",
     "typescript": "^4.9.4"
   },
